refactor backup code to allow us to plugin other backup engines
